port/enhancements: split monolithic enhancements.cpp into per-feature files
Each toggle/feature lived in one 595-line file. Move each to its own TU
matching the DisableStageHazards.cpp pattern already in this directory:
TapJump, HitboxView, StageClearFrozenWallpaper, CStickSmash, DPadJump,
AnalogRemap, UnlockCheats, Updater. Shared N64 button bitmasks go to
internal/buttons.h (used by CStickSmash and DPadJump).

CMake picks the new files up via the existing port/*.cpp GLOB; no build
config change required. Pure refactor — no behavior change. Boot-tested.
